requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.0.4
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-amazon
backoff
ratelimit
apache-airflow-providers-mysql
apache-airflow-providers-sftp
requests
oauthlib
requests_oauthlib
pyarrow
pandas
apache-airflow-providers-discord
google-auth
google-api-python-client
boto3
pysftp
psycopg2
mysql-connector-python